Moved in Youtube player package
